requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.2.0
beautifulsoup4==4.11.2
blis==0.7.9
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
confection==0.0.4
cymem==2.0.7
discord==2.2.2
discord.py==2.2.2
ffmpeg==1.4
frozenlist==1.3.3
google==3.0.0
gTTS==2.3.1
ibm-cloud-sdk-core==3.16.2
ibm-watson==6.1.0
idna==3.4
Jinja2==3.1.2
joblib==1.2.0
langcodes==3.3.0
MarkupSafe==2.1.2
multidict==6.0.4
murmurhash==1.0.9
numpy==1.24.2
openai==0.27.2
packaging==23.0
pathy==0.10.1
preshed==3.0.8
pycparser==2.21
pydantic==1.10.6
PyJWT==2.6.0
PyNaCl==1.5.0
python-dateutil==2.8.2
python-dotenv==1.0.0
regex==2022.10.31
requests==2.28.2
scipy==1.10.1
six==1.16.0
smart-open==6.3.0
sounddevice==0.4.6
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
thinc==8.1.9
tiktoken==0.3.2
tqdm==4.65.0
typer==0.7.0
typing_extensions==4.5.0
urllib3==1.26.15
wasabi==1.1.1
watson-developer-cloud==2.10.1
wavio==0.0.7
websocket-client==1.1.0
wikipedia==1.4.0
yarl==1.8.2
